refactor(sdk/analyze): consolidate percentile + hotspots_action helpers
Promote the byte-identical f64 percentile in claude_md/subagent_tree to a
single generic util::percentile<T: Copy + Default>. The u64 variant in
tool_output_bloat stays local (it scales p over 0..=100 and sorts internally).

Promote findings::hotspots_action to pub(crate) and drop the duplicate copy
in tool_call_patterns, matching how severity_from_usd is already shared.
